refactor: splite classes
